Fix/bug to laravel shift

+++ database/migrations/0001_01_01_000006_create_announcement_read_status_table.php
@@ -8,16 +8,18 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('announcement_read_status', function (Blueprint $table) {
-            $table->id();
-            $table->string('announcement_id', 50);
-            $table->string('user_id', 16);
-            $table->timestamp('read_at')->useCurrent();
+        if (!Schema::hasTable('announcement_read_status')) {
+            Schema::create('announcement_read_status', function (Blueprint $table) {
+                $table->id();
+                $table->string('announcement_id', 50);
+                $table->string('user_id', 16);
+                $table->timestamp('read_at')->useCurrent();
 
-            $table->unique(['announcement_id', 'user_id'], 'announce_user');
-            $table->foreign('announcement_id')->references('id')->on('announcements')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+                $table->unique(['announcement_id', 'user_id'], 'announce_user');
+                $table->foreign('announcement_id')->references('id')->on('announcements')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
